Reworked route string description.
The lower part of the input field can now be used as a scratch pad. The upper bold highlighted
section separated by an empty line is the current route description which is read and interpreted.
The lower gray description texts are a part of the scratch pad. These are saved on program exit but
ignored apart from that.

The function `Create flight plan` now reads the first paragraph in the input field, i.e. until
the first empty line.
'Load from flight plan' inserts the route description on top of the input field and leaves other
texts below intact.

Fixed issues where airways and waypoints were dropped in some cases.